Frontend/requirements.txt
streamlit
Pillow
requests
tensorflow==2.12.0
numpy